OfT class
